requirements.txt
Django==2.2.13
mysqlclient==2.0.3
celery==5.0
ipython==7.15
djangorestframework==3.12
django-filter==2.3
Markdown==3.2
Pygments==2.2
djangorestframework-simplejwt
Pillow==4.3
django-cors-middleware==1.5
django-crispy-forms==1.9
sorl-thumbnail==12.7.0

#dev
Faker==4.18.0
django-silk==4.0
pytest==6.1
pytest-cov==2.10
pytest-django==4.0
pytest-mock==3.3